chore: Fix initialization order

+++ src/libcommon/info/searchinfo.h
@@ -34,7 +34,7 @@ class SearchInfo {
         [[nodiscard]] NodeType type() const { return _type; }
         [[nodiscard]] const SyncPath &path() const { return _path; }
         [[nodiscard]] SyncTime lastModifiedTime() const { return _modifiedTime; }
-        [[nodiscard]] size_t size() const { return _size; }
+        [[nodiscard]] int64_t size() const { return _size; }
         [[nodiscard]] bool isAvailableLocally() const { return _isAvailableLocally; }
 
         void toDynamicStruct(Poco::DynamicStruct &dstruct) const;
@@ -50,7 +50,7 @@ class SearchInfo {
         SyncName _name;
         SyncPath _path;
         SyncTime _modifiedTime{0};
-        size_t _size{0};
+        int64_t _size{0};
         bool _isAvailableLocally{false};
         NodeType _type{NodeType::Unknown};
 };
